requirements.txt
PyMuPDF==1.23.6
Pillow
docx2txt
langchain
openai
tiktoken
unstructured
#unstructured[local-inference]
python-dotenv
chromadb
google-search-results
#dotenv